requirements.txt
arcade-ai~=1.0.5
python-dotenv~=1.0.1
streamlit~=1.43.1
